[platformio] Extract toolchain-agnostic PlatformIO library converter
Move the resolution/download/cache/manifest machinery out of espidf/component.py
into a shared esphome/platformio/library.py with a backend-parameterized
convert_libraries(). espidf/component.py becomes the ESP-IDF backend (CMakeLists
+ idf_component.yml emitter) over the shared module. No behavior change; existing
converter tests pass with patch targets repointed to the new module.
